requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
annotated-types==0.7.0
    # via pydantic
anyio==3.7.1
    # via
    #   fastapi
    #   starlette
cachetools==5.5.2
    # via google-auth
certifi==2025.4.26
    # via requests
charset-normalizer==3.4.2
    # via requests
click==8.1.8
    # via uvicorn
fastapi==0.104.1
    # via auto-lgtm (pyproject.toml)
google-api-core==2.24.2
    # via google-cloud-secret-manager
google-auth==2.40.1
    # via google-api-core
google-cloud-secret-manager==2.16.1
    # via auto-lgtm (pyproject.toml)
googleapis-common-protos==1.70.0
    # via
    #   google-api-core
    #   grpc-google-iam-v1
    #   grpcio-status
grpc-google-iam-v1==0.14.2
    # via google-cloud-secret-manager
grpcio==1.71.0
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
grpcio-status==1.62.3
    # via google-api-core
gunicorn==21.2.0
    # via auto-lgtm (pyproject.toml)
h11==0.16.0
    # via uvicorn
idna==3.10
    # via
    #   anyio
    #   requests
iniconfig==2.1.0
    # via pytest
loguru==0.7.3
    # via auto-lgtm (pyproject.toml)
markdown-it-py==3.0.0
    # via rich
mdurl==0.1.2
    # via markdown-it-py
packaging==25.0
    # via
    #   gunicorn
    #   pytest
pluggy==1.5.0
    # via pytest
proto-plus==1.26.1
    # via
    #   google-api-core
    #   google-cloud-secret-manager
protobuf==4.25.7
    # via
    #   google-api-core
    #   google-cloud-secret-manager
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
    #   proto-plus
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydantic==2.11.4
    # via fastapi
pydantic-core==2.33.2
    # via pydantic
pyfiglet==1.0.2
    # via auto-lgtm (pyproject.toml)
pygments==2.19.1
    # via rich
pytest==8.3.5
    # via auto-lgtm (pyproject.toml)
python-dotenv==1.0.0
    # via auto-lgtm (pyproject.toml)
requests==2.32.3
    # via
    #   auto-lgtm (pyproject.toml)
    #   google-api-core
rich==14.0.0
    # via auto-lgtm (pyproject.toml)
rsa==4.9.1
    # via google-auth
sniffio==1.3.1
    # via anyio
starlette==0.27.0
    # via fastapi
typing-extensions==4.13.2
    # via
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.0
    # via pydantic
urllib3==2.4.0
    # via requests
uvicorn==0.24.0
    # via auto-lgtm (pyproject.toml)
